peers: keep blocked list into memory till restart or pibd finish
